@@ -4,15 +4,16 @@ version: 0.5.1
 homepage: https://github.com/rinukkusu/spotify-dart
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: '>=2.12.0-259.9.beta <3.0.0'
 
 dependencies:
   http: ^0.13.1
   json_annotation: ^4.0.1
   oauth2: ^2.0.0
   pedantic: ^1.11.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   test: ^1.16.8
-  build_runner: ^1.12.2
+  build_runner: ^2.0.0
   json_serializable: ^4.1.0
