@@ -7,21 +7,21 @@ abstract class EndpointPaging extends EndpointBase {
   EndpointPaging(SpotifyApiBase api) : super(api);
 
   Pages<T> _getPages<T>(String path, ParserFunction<T> pageItemParser,
-          [String pageKey, ParserFunction<Object> pageContainerParser]) =>
+          [String? pageKey, ParserFunction<Object>? pageContainerParser]) =>
       Pages(_api, path, pageItemParser, pageKey, pageContainerParser);
 
   BundledPages _getBundledPages<T>(
           String path, Map<String, ParserFunction<T>> pageItemParsers,
-          [String pageKey, ParserFunction<Object> pageContainerParser]) =>
+          [String? pageKey, ParserFunction<Object>? pageContainerParser]) =>
       BundledPages(_api, path, pageItemParsers, pageKey, pageContainerParser);
 }
 
 class Page<T> {
   final Paging<T> _paging;
-  Iterable<T> _items;
-  Object _container;
+  Iterable<T>? _items;
+  Object? _container;
 
-  Page(this._paging, ParserFunction<T> pageItemParser, [Object pageContainer]) {
+  Page(this._paging, ParserFunction<T> pageItemParser, [Object? pageContainer]) {
     _items = _paging.itemsNative.map(pageItemParser);
     _container = pageContainer;
   }
@@ -33,13 +33,13 @@ class Page<T> {
   Paging<T> get metadata => _paging;
 
   /// The requested data
-  Iterable<T> get items => _items;
+  Iterable<T>? get items => _items;
 
   /// The object containing this page, if applicable
-  Object get container => _container;
+  Object? get container => _container;
 
-  bool get isLast => _paging.offset + _paging.limit >= _paging.total;
-  int get nextOffset => _paging.offset + _paging.limit;
+  bool get isLast => _paging.offset! + _paging.limit! >= _paging.total!;
+  int get nextOffset => _paging.offset! + _paging.limit!;
 }
 
 const defaultLimit = 20;
@@ -47,8 +47,8 @@ const defaultLimit = 20;
 abstract class _Pages<T> {
   final SpotifyApiBase _api;
   final String _path;
-  final String _pageKey;
-  final ParserFunction<Object> _pageContainerParser;
+  final String? _pageKey;
+  final ParserFunction<Object>? _pageContainerParser;
 
   _Pages(this._api, this._path, [this._pageKey, this._pageContainerParser]) {
     if (_pageKey != null && _pageContainerParser == null) {
@@ -71,12 +71,12 @@ class Pages<T> extends _Pages<Page<T>> {
   final List<Page<T>> _bufferedPages = [];
   bool _cancelled = false;
   Pages(SpotifyApiBase api, String path, this._pageParser,
-      [String pageKey, ParserFunction<Object> pageContainerMapper])
+      [String? pageKey, ParserFunction<Object>? pageContainerMapper])
       : super(api, path, pageKey, pageContainerMapper);
 
   Pages.fromPaging(SpotifyApiBase api, Paging<T> paging, this._pageParser,
-      [String pageKey, ParserFunction<Object> pageContainerMapper])
-      : super(api, Uri.parse(paging.href).path.substring(1), pageKey,
+      [String? pageKey, ParserFunction<Object>? pageContainerMapper])
+      : super(api, Uri.parse(paging.href!).path.substring(1), pageKey,
             pageContainerMapper) {
     _bufferedPages.add(Page<T>(paging, _pageParser));
   }
@@ -85,11 +85,11 @@ class Pages<T> extends _Pages<Page<T>> {
     return stream(limit)
         .map((page) => page.items)
         .toList()
-        .then((pages) => pages.expand((page) => page));
+        .then((pages) => pages.expand((page) => page!));
   }
 
   Stream<Page<T>> stream([int limit = defaultLimit]) {
-    StreamController<Page<T>> stream;
+    late StreamController<Page<T>> stream;
 
     void handlePageAndGetNext(Page<T> page) {
       if (_cancelled) {
@@ -149,17 +149,17 @@ class Pages<T> extends _Pages<Page<T>> {
       return Page<T>(paging, _pageParser);
     } else {
       var paging = Paging<T>.fromJson(map[_pageKey]);
-      var container = _pageContainerParser(map);
+      var container = _pageContainerParser!(map);
       return Page<T>(paging, _pageParser, container);
     }
   }
 }
 
 class BundledPages extends _Pages<List<Page<Object>>> {
-  final Map<String, ParserFunction<Object>> _pageMappers;
+  final Map<String, ParserFunction<Object?>> _pageMappers;
 
   BundledPages(SpotifyApiBase api, String path, this._pageMappers,
-      [String pageKey, ParserFunction<Object> pageContainerParser])
+      [String? pageKey, ParserFunction<Object>? pageContainerParser])
       : super(api, path, pageKey, pageContainerParser);
 
   @override
@@ -180,7 +180,7 @@ class BundledPages extends _Pages<List<Page<Object>>> {
         if (_pageContainerParser == null) {
           page = Page(paging, value);
         } else {
-          var container = _pageContainerParser(map[key]);
+          var container = _pageContainerParser!(map[key]);
           page = Page(paging, value, container);
         }
         pages.add(page);
