@@ -7,26 +7,26 @@ class Users extends EndpointPaging {
   @override
   String get _path => 'v1/users';
 
-  Me _me;
+  Me? _me;
 
-  Users(SpotifyApiBase api, Me me) : super(api){ 
+  Users(SpotifyApiBase api, Me? me) : super(api){ 
     _me = me;
   }
 
   @Deprecated('Use "SpotifyApi.me.get()"')
-  Future<User> me() => _me.get();
+  Future<User> me() => _me!.get();
 
   @Deprecated('Use "SpotifyApi.me.currentlyPlaying()"')
-  Future<Player> currentlyPlaying() => _me.currentlyPlaying();
+  Future<Player> currentlyPlaying() => _me!.currentlyPlaying();
 
   @Deprecated('Use "SpotifyApi.me.recentlyPlayed()"')
-  Future<Iterable<PlayHistory>> recentlyPlayed() => _me.recentlyPlayed();
+  Future<Iterable<PlayHistory>?> recentlyPlayed() => _me!.recentlyPlayed();
 
   @Deprecated('Use "SpotifyApi.me.topTracks()"')
-  Future<Iterable<Track>> topTracks() => _me.topTracks();
+  Future<Iterable<Track>> topTracks() => _me!.topTracks();
 
   @Deprecated('Use "SpotifyApi.me.devices()"')
-  Future<Iterable<Device>> devices() async => _me.devices();
+  Future<Iterable<Device>> devices() async => _me!.devices();
 
   Future<UserPublic> get(String userId) async {
     var jsonString = await _api._get('$_path/$userId');
