@@ -10,38 +10,38 @@ class Album extends AlbumSimple {
   factory Album.fromJson(Map<String, dynamic> json) => _$AlbumFromJson(json);
 
   /// The copyright statements of the album.
-  List<Copyright> copyrights;
+  List<Copyright>? copyrights;
 
   /// Known external IDs for the album.
   @JsonKey(name: 'external_ids')
-  ExternalIds externalIds;
+  ExternalIds? externalIds;
 
   /// A list of the genres the artist is associated with.
   /// For example:
   ///     "Prog Rock", "Post-Grunge".
   ///
   /// (If not yet classified, the array is empty.)
-  List<String> genres;
+  List<String>? genres;
 
   /// The label for the album.
-  String label;
+  String? label;
 
   /// The popularity of the artist. The value will be between 0 and 100, with 100
   /// being the most popular. The artist's popularity is calculated from the
   /// popularity of all the artist's tracks.
-  int popularity;
+  int? popularity;
 
   /// The date the album was first released, for example "1981-12-15".
   /// Depending on the precision, it might be shown as "1981" or "1981-12".
   @JsonKey(name: 'release_date')
   @override
-  String releaseDate;
+  String? releaseDate;
 
   /// The precision with which release_date value is known:
   ///     "year", "month", or "day".
   @JsonKey(name: 'release_date_precision')
   @override
-  String releaseDatePrecision;
+  String? releaseDatePrecision;
 }
 
 @JsonSerializable(createToJson: false)
@@ -52,65 +52,65 @@ class AlbumSimple extends Object {
       _$AlbumSimpleFromJson(json);
 
   /// Helper function that unwraps the items from the paging object.
-  static Iterable<TrackSimple> _extractTracksFromPage(Map<String, dynamic> json) {
+  static Iterable<TrackSimple>? _extractTracksFromPage(Map<String, dynamic>? json) {
     if (json == null) {
       return [];
     }
     return json.isEmpty
         ? []
-        : Paging.fromJson(json).itemsNative?.map(
+        : Paging.fromJson(json).itemsNative.map(
             (trackJson) => TrackSimple.fromJson(trackJson)
           );
   }
 
   /// The type of the album: one of "album", "single", or "compilation".
   @JsonKey(name: 'album_type')
-  String albumType;
+  String? albumType;
 
   /// The artists of the album. Each artist object includes a link in href to
   /// more detailed information about the artist.
-  List<ArtistSimple> artists;
+  List<ArtistSimple>? artists;
 
   /// The markets in which the album is available: ISO 3166-1 alpha-2 country
   /// codes. Note that an album is considered available in a market when at least
   /// 1 of its tracks is available in that market.
   @JsonKey(name: 'available_markets')
-  List<String> availableMarkets;
+  List<String>? availableMarkets;
 
   /// Known external URLs for this album.
   @JsonKey(name: 'external_urls')
-  ExternalUrls externalUrls;
+  ExternalUrls? externalUrls;
 
   /// A link to the Web API endpoint providing full details of the album.
-  String href;
+  String? href;
 
   /// The Spotify ID for the album.
-  String id;
+  String? id;
 
   /// The cover art for the album in various sizes, widest first.
-  List<Image> images;
+  List<Image>? images;
 
   /// The name of the album. In case of an album takedown, the value may be an
   /// empty string.
-  String name;
+  String? name;
 
   /// The date the album was first released, for example "1981-12-15".
   /// Depending on the precision, it might be shown as "1981" or "1981-12".
   @JsonKey(name: 'release_date')
-  String releaseDate;
+  String? releaseDate;
 
   /// The precision with which release_date value is known:
   ///     "year", "month", or "day".
   @JsonKey(name: 'release_date_precision')
-  String releaseDatePrecision;
+  String? releaseDatePrecision;
 
   /// The object type: "album"
-  String type;
+  String? type;
 
   /// The Spotify URI for the album.
-  String uri;
+  String? uri;
 
   /// The tracks of this album.
   @JsonKey(fromJson: _extractTracksFromPage)
-  Iterable<TrackSimple> tracks;
+  Iterable<TrackSimple>? tracks;
 }
