@@ -12,23 +12,23 @@ class Player extends Object {
   factory Player.fromJson(Map<String, dynamic> json) => _$PlayerFromJson(json);
 
   /// Unix Millisecond Timestamp when data was fetched
-  int timestamp;
+  int? timestamp;
 
   /// A [PlayerContext] Object. Can be [null].
-  PlayerContext context;
+  PlayerContext? context;
 
   /// Progress into the currently playing track. Can be [null].
-  int progress_ms;
+  int? progress_ms;
 
   /// The currently playing track. Can be [null].
-  Track item;
+  Track? item;
 
   /// The object type of the currently playing item. Can be one of track,
   /// episode, ad or unknown.
-  String currently_playing_type;
+  String? currently_playing_type;
 
   /// If something is currently playing.
-  bool is_playing;
+  bool? is_playing;
 }
 
 @JsonSerializable(createToJson: false)
@@ -39,14 +39,14 @@ class PlayerContext extends Object {
       _$PlayerContextFromJson(json);
 
   /// The external_urls of the context, or [null] if not available.
-  ExternalUrls external_urls;
+  ExternalUrls? external_urls;
 
   /// The href of the context, or [null] if not available.
-  String href;
+  String? href;
 
   /// The object type of the item’s context. Can be one of album, artist or playlist.
-  String type;
+  String? type;
 
   /// The uri of the context.
-  String uri;
+  String? uri;
 }
