@@ -12,27 +12,27 @@ class Playlist extends Object implements PlaylistSimple {
 
   /// true if the owner allows other users to modify the playlist.
   @override
-  bool collaborative;
+  bool? collaborative;
 
   /// The playlist description. Only returned for modified, verified playlists,
   /// otherwise [null].
-  String description;
+  String? description;
 
   /// Known external URLs for this playlist.
   @JsonKey(name: 'external_urls')
   @override
-  ExternalUrls externalUrls;
+  ExternalUrls? externalUrls;
 
   /// Information about the followers of the playlist.
-  Followers followers;
+  Followers? followers;
 
   /// A link to the Web API endpoint providing full details of the playlist.
   @override
-  String href;
+  String? href;
 
   /// The Spotify ID for the playlist.
   @override
-  String id;
+  String? id;
 
   /// Images for the playlist. The array may be empty or contain up to three
   /// images. The images are returned by size in descending order. See Working
@@ -41,43 +41,43 @@ class Playlist extends Object implements PlaylistSimple {
   /// Note: If returned, the source URL for the image (url) is temporary and
   /// will expire in less than a day.
   @override
-  List<Image> images;
+  List<Image>? images;
 
   /// The name of the playlist.
   @override
-  String name;
+  String? name;
 
   /// The user who owns the playlist
   @override
-  User owner;
+  User? owner;
 
   /// The playlist's public/private status: [true] the playlist is public,
   /// [false] the playlist is private, null the playlist status is not relevant.
   /// For more about public/private status, see Working with Playlists.
   @override
-  bool public;
+  bool? public;
 
   /// The version identifier for the current playlist. Can be supplied in other
   /// requests to target a specific playlist version
   @JsonKey(name: 'snapshot_id')
   @override
-  String snapshotId;
+  String? snapshotId;
 
   /// Use [Playlist.tracks]
   @override
   @JsonKey(ignore: true)
-  TracksLink tracksLink;
+  TracksLink? tracksLink;
 
   /// Information about the tracks of the playlist.
-  Paging<Track> tracks;
+  Paging<Track>? tracks;
 
   /// The object type: "playlist"
   @override
-  String type;
+  String? type;
 
   /// The Spotify URI for the playlist.
   @override
-  String uri;
+  String? uri;
 }
 
 @JsonSerializable(createToJson: false)
@@ -87,17 +87,17 @@ class PlaylistSimple extends Object {
       _$PlaylistSimpleFromJson(json);
 
   /// true if the owner allows other users to modify the playlist.
-  bool collaborative;
+  bool? collaborative;
 
   /// Known external URLs for this playlist.
   @JsonKey(name: 'external_urls')
-  ExternalUrls externalUrls;
+  ExternalUrls? externalUrls;
 
   /// A link to the Web API endpoint providing full details of the playlist.
-  String href;
+  String? href;
 
   /// The Spotify ID for the playlist.
-  String id;
+  String? id;
 
   /// Images for the playlist. The array may be empty or contain up to three
   /// images. The images are returned by size in descending order. See Working
@@ -105,35 +105,35 @@ class PlaylistSimple extends Object {
   ///
   /// Note: If returned, the source URL for the image (url) is temporary and
   /// will expire in less than a day.
-  List<Image> images;
+  List<Image>? images;
 
   /// The name of the playlist.
-  String name;
+  String? name;
 
   /// The user who owns the playlist
-  User owner;
+  User? owner;
 
   /// The playlist's public/private status: [true] the playlist is public,
   /// [false] the playlist is private, null the playlist status is not relevant.
   /// For more about public/private status, see Working with Playlists.
-  bool public;
+  bool? public;
 
   /// The version identifier for the current playlist. Can be supplied in other
   /// requests to target a specific playlist version
   @JsonKey(name: 'snapshot_id')
-  String snapshotId;
+  String? snapshotId;
 
   /// A collection containing a link (href) to the Web API endpoint where full
   /// details of the playlist's tracks can be retrieved, along with the total
   /// number of tracks in the playlist.
   @JsonKey(name: 'tracks')
-  TracksLink tracksLink;
+  TracksLink? tracksLink;
 
   /// The object type: "playlist"
-  String type;
+  String? type;
 
   /// The Spotify URI for the playlist.
-  String uri;
+  String? uri;
 }
 
 @JsonSerializable(createToJson: false)
@@ -143,7 +143,7 @@ class PlaylistsFeatured extends Object {
       _$PlaylistsFeaturedFromJson(json);
 
   /// The message of the day for Spotify's featured playlists
-  String message;
+  String? message;
 }
 
 @JsonSerializable(createToJson: false)
@@ -155,17 +155,17 @@ class PlaylistTrack extends Object {
   /// The date and time the track was added.
   /// Note that some very old playlists may return [null] in this field.
   @JsonKey(name: 'added_at')
-  DateTime addedAt;
+  DateTime? addedAt;
 
   /// The Spotify user who added the track.
   /// Note that some very old playlists may return [null] in this field.
   @JsonKey(name: 'added_by')
-  UserPublic addedBy;
+  UserPublic? addedBy;
 
   /// Whether this track is a local file or not.
   @JsonKey(name: 'is_local')
-  bool isLocal;
+  bool? isLocal;
 
   /// Information about the track
-  Track track;
+  Track? track;
 }
