@@ -11,62 +11,62 @@ class Track extends Object implements TrackSimple {
 
   /// The album on which the track appears. The album object includes a link
   /// in [href] to full information about the album.
-  AlbumSimple album;
+  AlbumSimple? album;
 
   /// The artists who performed the track. Each artist object includes a link
   /// in [href] to more detailed information about the artist.
   @override
-  List<Artist> artists;
+  List<Artist>? artists;
 
   /// A list of the countries in which the track can be played, identified by
   /// their ISO 3166-1 alpha-2 code.
   @JsonKey(name: 'available_markets')
   @override
-  List<String> availableMarkets;
+  List<String>? availableMarkets;
 
   /// The disc number
   /// (usually [1] unless the album consists of more than one disc)
   @JsonKey(name: 'disc_number')
   @override
-  int discNumber;
+  int? discNumber;
 
   /// The track length in milliseconds.
   @JsonKey(name: 'duration_ms')
   @override
-  int durationMs;
+  int? durationMs;
 
   /// The track length
   @JsonKey(ignore: true)
   @override
-  Duration get duration => Duration(milliseconds: durationMs);
+  Duration get duration => Duration(milliseconds: durationMs!);
 
   /// Whether or not the track has explicit lyrics
   /// ([true] = yes it does; [false] = no it does not OR unknown).
   @override
-  bool explicit;
+  bool? explicit;
 
   /// Known external IDs for this track.
   @JsonKey(name: 'external_ids')
-  ExternalIds externalIds;
+  ExternalIds? externalIds;
 
   /// Known external URLs for this track.
   @JsonKey(name: 'external_urls')
   @override
-  ExternalUrls externalUrls;
+  ExternalUrls? externalUrls;
 
   /// A link to the Web API endpoint providing full details of the track.
   @override
-  String href;
+  String? href;
 
   /// The Spotify ID for the track.
   @override
-  String id;
+  String? id;
 
   /// Part of the response when Track Relinking is applied. If true, the track
   /// is playable in the given market. Otherwise false.
   @JsonKey(name: 'is_playable')
   @override
-  bool isPlayable;
+  bool? isPlayable;
 
   /// Part of the response when Track Relinking is applied and is only part of
   /// the response if the track linking, in fact, exists. The requested track
@@ -74,11 +74,11 @@ class Track extends Object implements TrackSimple {
   /// object contains information about the originally requested track.
   @JsonKey(name: 'linked_from')
   @override
-  TrackLink linkedFrom;
+  TrackLink? linkedFrom;
 
   /// The name of the track.
   @override
-  String name;
+  String? name;
 
   /// The popularity of the track. The value will be between 0 and 100, with 100
   /// being the most popular.
@@ -94,27 +94,27 @@ class Track extends Object implements TrackSimple {
   /// independently. Artist and album popularity is derived mathematically from
   /// track popularity. Note that the popularity value may lag actual popularity
   /// by a few days: the value is not updated in real time.
-  int popularity;
+  int? popularity;
 
   /// A URL to a 30 second preview (MP3 format) of the track. [null] if not
   /// available.
   @JsonKey(name: 'preview_url')
   @override
-  String previewUrl;
+  String? previewUrl;
 
   /// The number of the track. If an album has several discs, the track number
   /// is the number on the specified disc.
   @JsonKey(name: 'track_number')
   @override
-  int trackNumber;
+  int? trackNumber;
 
   /// The object type: "track".
   @override
-  String type;
+  String? type;
 
   /// The Spotify URI for the track.
   @override
-  String uri;
+  String? uri;
 }
 
 @JsonSerializable(createToJson: false)
@@ -126,70 +126,70 @@ class TrackSimple extends Object {
 
   /// The artists who performed the track. Each artist object includes a link
   /// in [href] to more detailed information about the artist.
-  List<Artist> artists;
+  List<Artist>? artists;
 
   /// A list of the countries in which the track can be played, identified by
   /// their ISO 3166-1 alpha-2 code.
   @JsonKey(name: 'available_markets')
-  List<String> availableMarkets;
+  List<String>? availableMarkets;
 
   /// The disc number
   /// (usually [1] unless the album consists of more than one disc)
   @JsonKey(name: 'disc_number')
-  int discNumber;
+  int? discNumber;
 
   /// The track length in milliseconds.
   @JsonKey(name: 'duration_ms')
-  int durationMs;
+  int? durationMs;
 
   /// The track length
   @JsonKey(ignore: true)
-  Duration get duration => Duration(milliseconds: durationMs);
+  Duration get duration => Duration(milliseconds: durationMs!);
 
   /// Whether or not the track has explicit lyrics
   /// ([true] = yes it does; [false] = no it does not OR unknown).
-  bool explicit;
+  bool? explicit;
 
   /// Known external URLs for this track.
   @JsonKey(name: 'external_urls')
-  ExternalUrls externalUrls;
+  ExternalUrls? externalUrls;
 
   /// A link to the Web API endpoint providing full details of the track.
-  String href;
+  String? href;
 
   /// The Spotify ID for the track.
-  String id;
+  String? id;
 
   /// Part of the response when Track Relinking is applied. If true, the track
   /// is playable in the given market. Otherwise false.
   @JsonKey(name: 'is_playable')
-  bool isPlayable;
+  bool? isPlayable;
 
   /// Part of the response when Track Relinking is applied and is only part of
   /// the response if the track linking, in fact, exists. The requested track
   /// has been replaced with a different track. The track in the linked_from
   /// object contains information about the originally requested track.
   @JsonKey(name: 'linked_from')
-  TrackLink linkedFrom;
+  TrackLink? linkedFrom;
 
   /// The name of the track.
-  String name;
+  String? name;
 
   /// A URL to a 30 second preview (MP3 format) of the track. [null] if not
   /// available.
   @JsonKey(name: 'preview_url')
-  String previewUrl;
+  String? previewUrl;
 
   /// The number of the track. If an album has several discs, the track number
   /// is the number on the specified disc.
   @JsonKey(name: 'track_number')
-  int trackNumber;
+  int? trackNumber;
 
   /// The object type: "track".
-  String type;
+  String? type;
 
   /// The Spotify URI for the track.
-  String uri;
+  String? uri;
 }
 
 /// A song saved in a Spotify user’s “Your Music” library
@@ -201,10 +201,10 @@ class TrackSaved extends Object {
 
   /// The date and time the track was saved.
   @JsonKey(name: 'added_at')
-  DateTime addedAt;
+  DateTime? addedAt;
 
   /// Information about the track.
-  Track track;
+  Track? track;
 }
 
 @JsonSerializable(createToJson: false)
@@ -215,19 +215,19 @@ class TrackLink extends Object {
 
   /// Known external URLs for this track.
   @JsonKey(name: 'external_urls')
-  Map<String, String> externalUrls;
+  Map<String, String>? externalUrls;
 
   /// A link to the Web API endpoint providing full details of the track.
-  String href;
+  String? href;
 
   /// The Spotify ID for the track.
-  String id;
+  String? id;
 
   /// The object type: "track".
-  String type;
+  String? type;
 
   /// The Spotify URI for the track.
-  String uri;
+  String? uri;
 }
 
 @JsonSerializable(createToJson: false)
@@ -238,8 +238,8 @@ class TracksLink extends Object {
 
   /// A link to the Web API endpoint where full details of the playlist's
   /// tracks can be retrieved
-  String href;
+  String? href;
 
   /// Total number of tracks in the playlist
-  int total;
+  int? total;
 }
