@@ -8,39 +8,39 @@ class SpotifyApiCredentials {
   ///
   /// Spotify issues each client a separate client identifier and secret,
   /// which allows the server to tell which client is accessing it.
-  String clientId;
+  String? clientId;
 
   /// The client secret for this Spotify client.
   ///
   /// Spotify issues each client a separate client identifier and secret,
   /// which allows the server to tell which client is accessing it.
-  String clientSecret;
+  String? clientSecret;
 
   /// The token that is sent to Spotify to prove the authorization of a client.
-  String accessToken;
+  String? accessToken;
 
   /// The token that is sent to Spotify to refresh the credentials.
   ///
   /// This may be `null`, indicating that the credentials can't be refreshed.
-  String refreshToken;
+  String? refreshToken;
 
   /// The URL of the Spotify endpoint that's used to refresh the credentials.
   ///
   /// This may be `null`, indicating that the credentials can't be refreshed.
-  Uri tokenEndpoint;
+  Uri? tokenEndpoint;
 
   /// The specific permissions being requested from Spotify.
   ///
   /// See https://developer.spotify.com/documentation/general/guides/scopes/
   /// for a full list of available scopes.
-  List<String> scopes;
+  List<String>? scopes;
 
   /// The date at which these credentials will expire, stored in the user's
   /// local time.
   ///
   /// This is likely to be a few seconds earlier than the server's idea of the
   /// expiration date.
-  DateTime expiration;
+  DateTime? expiration;
 
   SpotifyApiCredentials(
     this.clientId,
@@ -55,13 +55,11 @@ class SpotifyApiCredentials {
     clientId = client.identifier;
     clientSecret = client.secret;
 
-    if (client.credentials != null) {
-      accessToken = client.credentials.accessToken;
-      refreshToken = client.credentials.refreshToken;
-      tokenEndpoint = client.credentials.tokenEndpoint;
-      scopes = client.credentials.scopes;
-      expiration = client.credentials.expiration;
-    }
+    accessToken = client.credentials.accessToken;
+    refreshToken = client.credentials.refreshToken;
+    tokenEndpoint = client.credentials.tokenEndpoint;
+    scopes = client.credentials.scopes;
+    expiration = client.credentials.expiration;
   }
 
   /// Whether or not these credentials have expired.
@@ -70,7 +68,7 @@ class SpotifyApiCredentials {
   /// called. However, since the client's expiration date is kept a few seconds
   /// earlier than the server's, there should be enough leeway to rely on this.
   bool get isExpired =>
-      expiration != null && DateTime.now().isAfter(expiration);
+      expiration != null && DateTime.now().isAfter(expiration!);
 
   /// Whether it's possible to refresh these credentials.
   bool get canRefresh => refreshToken != null && tokenEndpoint != null;
@@ -89,7 +87,7 @@ class SpotifyApiCredentials {
       expiration != null;
 
   oauth2.Credentials _toOauth2Credentials() => oauth2.Credentials(
-        accessToken,
+        accessToken!,
         refreshToken: refreshToken,
         tokenEndpoint: tokenEndpoint,
         scopes: scopes,
