@@ -11,19 +11,19 @@ Future main() async {
 
   group('Albums', () {
     test('get', () async {
-      var album = await spotify.albums.get('4aawyAB9vmqN3uQ7FjRGTy');
+      var album = await spotify.albums!.get('4aawyAB9vmqN3uQ7FjRGTy');
 
       expect(album.albumType, 'album');
       expect(album.id, '4aawyAB9vmqN3uQ7FjRGTy');
-      expect(album.images.length, 3);
+      expect(album.images!.length, 3);
     });
 
     test('album tracks', () async {
-      var album = await spotify.albums.get('4aawyAB9vmqN3uQ7FjRGTy');
-      var items = album.tracks;
+      var album = await spotify.albums!.get('4aawyAB9vmqN3uQ7FjRGTy');
+      var items = album.tracks!;
       
       expect(items.length, 3);
-      var tracks = album.tracks;
+      var tracks = album.tracks!;
 
       var trackOne = tracks.first;
       expect(trackOne.discNumber, 1);
@@ -40,17 +40,17 @@ Future main() async {
       expect(trackOne.name, 'Global Warming');
 
       expect(trackOne.externalUrls != null, true);
-      expect(trackOne.externalUrls.spotify,
+      expect(trackOne.externalUrls!.spotify,
           'https://open.spotify.com/track/6OmhkSOpvYBokMKQxpIGx2');
 
-      var artists = trackOne.artists;
+      var artists = trackOne.artists!;
       expect(artists.length, 2);
       expect(artists[0].name, 'Pitbull');
       expect(artists[1].name, 'Sensato');
     });
 
     test('list', () async {
-      var albums = await spotify.albums
+      var albums = await spotify.albums!
           .list(['4aawyAB9vmqN3uQ7FjRGTy', '4aawyAB9vmqN3uQ7FjRGTy']);
 
       expect(albums.length, 2);
@@ -59,14 +59,14 @@ Future main() async {
 
   group('Artists', () {
     test('get', () async {
-      var artist = await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+      var artist = await spotify.artists!.get('0TnOYISbd1XYRBk9myaseg');
       expect(artist.type, 'artist');
       expect(artist.id, '0TnOYISbd1XYRBk9myaseg');
-      expect(artist.images.length, 3);
+      expect(artist.images!.length, 3);
     });
 
     test('list', () async {
-      var artists = await spotify.artists
+      var artists = await spotify.artists!
           .list(['0TnOYISbd1XYRBk9myaseg', '0TnOYISbd1XYRBk9myaseg']);
 
       expect(artists.length, 2);
@@ -75,22 +75,22 @@ Future main() async {
     test('getError', () async {
       spotify.mockHttpErrors =
           [MockHttpError(statusCode: 401, message: 'Bad Request')].iterator;
-      SpotifyException ex;
+      SpotifyException? ex;
       try {
-        await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+        await spotify.artists!.get('0TnOYISbd1XYRBk9myaseg');
       } catch (e) {
         expect(e, isA<SpotifyException>());
-        ex = e;
+        ex = e as SpotifyException;
       }
       expect(ex, isNotNull);
-      expect(ex.status, 401);
+      expect(ex!.status, 401);
       expect(ex.message, 'Bad Request');
     });
   });
 
   group('Shows', () {
     test('get', () async {
-      var show = await spotify.shows.get('4AlxqGkkrqe0mfIx3Mi7Xt');
+      var show = await spotify.shows!.get('4AlxqGkkrqe0mfIx3Mi7Xt');
 
       expect(show.type, 'show');
       expect(show.id, '4AlxqGkkrqe0mfIx3Mi7Xt');
@@ -98,7 +98,7 @@ Future main() async {
     });
 
     test('list', () async {
-      var shows = await spotify.shows
+      var shows = await spotify.shows!
           .list(['4AlxqGkkrqe0mfIx3Mi7Xt', '4AlxqGkkrqe0mfIx3Mi7Xt']);
 
       expect(shows.length, 2);
@@ -107,8 +107,8 @@ Future main() async {
 
   group('Show episodes', () {
     test('list', () async {
-      var episodes = spotify.shows.episodes('4AlxqGkkrqe0mfIx3Mi7Xt');
-      var firstEpisode = (await episodes.first()).items.first;
+      var episodes = spotify.shows!.episodes('4AlxqGkkrqe0mfIx3Mi7Xt');
+      Episode firstEpisode = (await episodes.first()).items!.first;
 
       expect(firstEpisode.type, 'episode');
       expect(firstEpisode.explicit, false);
@@ -117,35 +117,35 @@ Future main() async {
 
   group('Search', () {
     test('get', () async {
-      var searchResult = await spotify.search.get('metallica').first();
+      List<Page<Object>> searchResult = await spotify.search!.get('metallica').first();
       expect(searchResult.length, 2);
     });
 
     test('getError', () async {
       spotify.mockHttpErrors =
           [MockHttpError(statusCode: 401, message: 'Bad Request')].iterator;
-      SpotifyException ex;
+      SpotifyException? ex;
       try {
-        await spotify.search.get('metallica').first();
+        await spotify.search!.get('metallica').first();
       } catch (e) {
         expect(e, isA<SpotifyException>());
-        ex = e;
+        ex = e as SpotifyException;
       }
       expect(ex, isNotNull);
-      expect(ex.status, 401);
+      expect(ex!.status, 401);
       expect(ex.message, 'Bad Request');
     });
   });
 
   group('User', () {
     test('currentlyPlaying', () async {
-      var result = await spotify.me.currentlyPlaying();
+      var result = await spotify.me!.currentlyPlaying();
 
-      expect(result.item.name, 'So Voce');
+      expect(result.item!.name, 'So Voce');
     });
 
     test('devices', () async {
-      var result = await spotify.me.devices();
+      var result = await spotify.me!.devices();
       expect(result.length, 1);
       expect(result.first.id, '5fbb3ba6aa454b5534c4ba43a8c7e8e45a63ad0e');
       expect(result.first.isActive, true);
@@ -159,22 +159,22 @@ Future main() async {
     test('recentlyPlayed', () async {
       // the parameters don't do anything. They are just dummies
       var result =
-          await spotify.me.recentlyPlayed(limit: 3, before: DateTime.now());
+          await (spotify.me!.recentlyPlayed(limit: 3, before: DateTime.now()) as FutureOr<Iterable<PlayHistory>>);
       expect(result.length, 2);
       var first = result.first;
       expect(first.track != null, true);
 
       // just testing some sample attributes
-      var firstTrack = first.track;
+      var firstTrack = first.track!;
       expect(firstTrack.durationMs, 108546);
       expect(firstTrack.explicit, false);
       expect(firstTrack.id, '2gNfxysfBRfl9Lvi9T3v6R');
-      expect(firstTrack.artists.length, 1);
-      expect(firstTrack.artists.first.name, 'Tame Impala');
+      expect(firstTrack.artists!.length, 1);
+      expect(firstTrack.artists!.first.name, 'Tame Impala');
 
       var second = result.last;
       expect(second.playedAt, DateTime.tryParse('2016-12-13T20:42:17.016Z'));
-      expect(second.context.uri, 'spotify:artist:5INjqkS1o8h1imAzPqGZBb');
+      expect(second.context!.uri, 'spotify:artist:5INjqkS1o8h1imAzPqGZBb');
     });
   });
 
@@ -186,9 +186,9 @@ Future main() async {
       expect(result.clientSecret, 'clientSecret');
       expect(result.accessToken, 'accessToken');
       expect(result.refreshToken, 'refreshToken');
-      expect(result.tokenEndpoint.path, 'tokenEndpoint.com');
-      expect(result.scopes.length, 2);
-      expect(result.expiration.millisecondsSinceEpoch, 8000);
+      expect(result.tokenEndpoint!.path, 'tokenEndpoint.com');
+      expect(result.scopes!.length, 2);
+      expect(result.expiration!.millisecondsSinceEpoch, 8000);
       expect(result.canRefresh, true);
       expect(result.isExpired, true);
     });
@@ -201,10 +201,10 @@ Future main() async {
               statusCode: 429,
               message: 'API Rate exceeded',
               headers: {'retry-after': '1'})).iterator;
-      var artist = await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+      var artist = await spotify.artists!.get('0TnOYISbd1XYRBk9myaseg');
       expect(artist.type, 'artist');
       expect(artist.id, '0TnOYISbd1XYRBk9myaseg');
-      expect(artist.images.length, 3);
+      expect(artist.images!.length, 3);
     });
     test('apiRateErrorFail', () async {
       spotify.mockHttpErrors = List.generate(
@@ -213,15 +213,15 @@ Future main() async {
               statusCode: 429,
               message: 'API Rate exceeded',
               headers: {'retry-after': '1'})).iterator;
-      ApiRateException ex;
+      ApiRateException? ex;
       try {
-        await spotify.artists.get('0TnOYISbd1XYRBk9myaseg');
+        await spotify.artists!.get('0TnOYISbd1XYRBk9myaseg');
       } catch (e) {
         expect(e, isA<ApiRateException>());
-        ex = e;
+        ex = e as ApiRateException;
       }
       expect(ex, isNotNull);
-      expect(ex.status, 429);
+      expect(ex!.status, 429);
     });
   });
 }
