@@ -13,36 +13,36 @@ void main() async {
   var spotify = SpotifyApi(credentials);
 
   print('\nPodcast:');
-  var podcast = await spotify.shows.get('4AlxqGkkrqe0mfIx3Mi7Xt');
+  var podcast = await spotify.shows!.get('4AlxqGkkrqe0mfIx3Mi7Xt');
   print(podcast.name);
 
   print('\nPodcast episode:');
-  var episodes = spotify.shows.episodes('4AlxqGkkrqe0mfIx3Mi7Xt');
-  var firstEpisode = (await episodes.first()).items.first;
+  var episodes = spotify.shows!.episodes('4AlxqGkkrqe0mfIx3Mi7Xt');
+  Episode firstEpisode = (await episodes.first()).items!.first;
   print(firstEpisode.name);
 
   print('Artists:');
-  var artists = await spotify.artists.list(['0OdUWJ0sBjDrqHygGUXeCF']);
+  var artists = await spotify.artists!.list(['0OdUWJ0sBjDrqHygGUXeCF']);
   artists.forEach((x) => print(x.name));
 
   print('\nAlbum:');
-  var album = await spotify.albums.get('2Hog1V8mdTWKhCYqI5paph');
+  var album = await spotify.albums!.get('2Hog1V8mdTWKhCYqI5paph');
   print(album.name);
 
   print('\nAlbum Tracks:');
-  var tracks = await spotify.albums.getTracks(album.id).all();
+  var tracks = await spotify.albums!.getTracks(album.id).all();
   tracks.forEach((track) {
     print(track.name);
   });
 
   print('\nFeatured Playlist:');
-  var featuredPlaylists = await spotify.playlists.featured.all();
+  var featuredPlaylists = await spotify.playlists!.featured.all();
   featuredPlaylists.forEach((playlist) {
     print(playlist.name);
   });
 
   print("\nSearching for \'Metallica\':");
-  var search = await spotify.search
+  List<Page<Object>> search = await spotify.search!
       .get('metallica')
       .first(2)
       .catchError((err) => print((err as SpotifyException).message));
@@ -50,20 +50,20 @@ void main() async {
     return;
   }
   search.forEach((pages) {
-    pages.items.forEach((item) {
+    pages.items!.forEach((item) {
       if (item is PlaylistSimple) {
         print('Playlist: \n'
             'id: ${item.id}\n'
             'name: ${item.name}:\n'
             'collaborative: ${item.collaborative}\n'
             'href: ${item.href}\n'
-            'trackslink: ${item.tracksLink.href}\n'
+            'trackslink: ${item.tracksLink!.href}\n'
             'owner: ${item.owner}\n'
             'public: ${item.owner}\n'
             'snapshotId: ${item.snapshotId}\n'
             'type: ${item.type}\n'
             'uri: ${item.uri}\n'
-            'images: ${item.images.length}\n'
+            'images: ${item.images!.length}\n'
             '-------------------------------');
       }
       if (item is Artist) {
@@ -83,8 +83,8 @@ void main() async {
             'type: ${item.type}\n'
             'uri: ${item.uri}\n'
             'isPlayable: ${item.isPlayable}\n'
-            'artists: ${item.artists.length}\n'
-            'availableMarkets: ${item.availableMarkets.length}\n'
+            'artists: ${item.artists!.length}\n'
+            'availableMarkets: ${item.availableMarkets!.length}\n'
             'discNumber: ${item.discNumber}\n'
             'trackNumber: ${item.trackNumber}\n'
             'explicit: ${item.explicit}\n'
@@ -98,9 +98,9 @@ void main() async {
             'type: ${item.type}\n'
             'uri: ${item.uri}\n'
             'albumType: ${item.albumType}\n'
-            'artists: ${item.artists.length}\n'
-            'availableMarkets: ${item.availableMarkets.length}\n'
-            'images: ${item.images.length}\n'
+            'artists: ${item.artists!.length}\n'
+            'availableMarkets: ${item.availableMarkets!.length}\n'
+            'images: ${item.images!.length}\n'
             'releaseDate: ${item.releaseDate}\n'
             'releaseDatePrecision: ${item.releaseDatePrecision}\n'
             '-------------------------------');
@@ -109,7 +109,7 @@ void main() async {
   });
 
   var relatedArtists =
-      await spotify.artists.relatedArtists('0OdUWJ0sBjDrqHygGUXeCF');
+      await spotify.artists!.relatedArtists('0OdUWJ0sBjDrqHygGUXeCF');
   print('\nRelated Artists: ${relatedArtists.length}');
 
   credentials = await spotify.getCredentials();
