@@ -4,9 +4,11 @@
 part of spotify;
 
 class SpotifyApi extends SpotifyApiBase {
-  SpotifyApi(SpotifyApiCredentials credentials, {Function(SpotifyApiCredentials) onCredentialsRefreshed})
-      : super(credentials, http.Client(), onCredentialsRefreshed);
-  
+  SpotifyApi(SpotifyApiCredentials credentials,
+      {Function(SpotifyApiCredentials)? onCredentialsRefreshed})
+      : super(credentials, http.Client() as http.BaseClient?,
+            onCredentialsRefreshed);
+
   SpotifyApi.fromClient(FutureOr<oauth2.Client> client)
       : super.fromClient(client);
 
@@ -15,9 +17,9 @@ class SpotifyApi extends SpotifyApiBase {
       : super.fromAuthCodeGrant(grant, responseUri);
 
   static oauth2.AuthorizationCodeGrant authorizationCodeGrant(
-      SpotifyApiCredentials credentials, {Function(SpotifyApiCredentials) onCredentialsRefreshed}) {
-    return SpotifyApiBase.authorizationCodeGrant(credentials, http.Client(), onCredentialsRefreshed);
+      SpotifyApiCredentials credentials,
+      {Function(SpotifyApiCredentials)? onCredentialsRefreshed}) {
+    return SpotifyApiBase.authorizationCodeGrant(
+        credentials, http.Client() as http.BaseClient, onCredentialsRefreshed);
   }
-
-   
 }
